Added the signuppage, in theory this works, as I reused the logic for now from the last versions
