+++ src/components/Auth/AuthComponents/index.js
@@ -1,8 +1,7 @@
-// src/components/Auth/index.js
 export { FormInput } from './FormInput';
 export { AuthButton } from './AuthButton';
 export { GoogleAuthButton } from './GoogleAuthButton';
 export { ErrorMessage } from './ErrorMessage';
 export { AuthLink } from './AuthLink';
-export { default as LoginPage } from '../LoginPage';
+//export { default as LoginPage } from '../LoginPage';
 //export { default as SignUpPage } from './SignUpPage';
