--- src/components/firebase-test/Foro.jsx
@@ -1,10 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-import React from "react";
-function Foro() {
-    return (
-        <div>
-            <h1>Foro</h1>
-        </div>
-    );
-}
-export default Foro;
