--- src/components/firebase-test/Home.jsx
@@ -1,10 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-import React from "react";
-function Home() {
-    return (
-        <div>
-            <h1>Home</h1>
-        </div>
-    );
-}
-export default Home;
