--- src/components/firebase-test/Login.jsx
@@ -1,10 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-import React from "react";
-function Login() {
-    return (
-        <div>
-            <h1>Login</h1>
-        </div>
-    );
-}
-export default Login;
