--- src/components/firebase-test/Register.jsx
@@ -1,10 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-import React from "react";
-function Register() {
-    return (
-        <div>
-            <h1>Register</h1>
-        </div>
-    );
-}
-export default Register;
