--- src/components/firebase-test/StorageTest/StorageTest.css
@@ -1,55 +1,0 @@
-body {
-    font-family: 'Arial', sans-serif;
-    background-color: #021200;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    margin: 0;
-}
-
-.storage-container {
-    background-color: #fff;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    width: 300px;
-    text-align: center;
-}
-
-h2 {
-    margin-bottom: 20px;
-    color: #ffffff;
-}
-
-input[type="file"] {
-    width: calc(100% - 20px);
-    padding: 10px;
-    margin: 5px 0;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    font-size: 16px;
-}
-
-button.upload-button {
-    width: 100%;
-    padding: 15px;
-    margin: 10px 0;
-    background-color: #007bff;
-    color: #fff;
-    border: none;
-    border-radius: 6px;
-    cursor: pointer;
-    font-size: 18px;
-    transition: background-color 0.3s ease;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-button.upload-button:hover {
-    background-color: #0056b3;
-}
-
-button.upload-button:active {
-    background-color: #003f87;
-    transform: translateY(2px);
-}
